requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.23.2

# Data handling
numpy==1.27.5
pickle5==0.0.13

# FAISS for vector search
faiss-cpu==1.7.4

# Transformers + HuggingFace
transformers==4.42.2
torch==2.1.1+cpu
sentence-transformers==2.2.2
langchain-community==0.1.1

# Optional PDF generation (if you enable generate_pdf)
reportlab==4.0.0

# Embeddings
langchain_community==0.1.1
